separated table files and functions
