+++ src/components/PeopleTable/types.ts
@@ -1,3 +1,6 @@
+import { ContextDataType } from '../../utils/context/types';
+import { PeoplePageStateType } from '../PeoplePage';
+
 export enum Columns {
   name,
   sex,
@@ -11,3 +14,9 @@ export type SearchParams = {
   key: string;
   value: string;
 }[];
+
+export type FetchDBParams = {
+  contextData: ContextDataType;
+  setPeoplePageState: React.Dispatch<React.SetStateAction<PeoplePageStateType>>;
+  peoplePageState: PeoplePageStateType;
+};
